feat(ui): change set passcode to shared component
