+++ src/ui/components/CreatePasscodeModule/CreatePasscodeModule.scss
@@ -1,0 +1,30 @@
+.create-passcode-module {
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  margin: 0 !important;
+  justify-content: space-evenly;
+
+  > * {
+    flex: 0 1 auto;
+  }
+
+  & > .passcode-module-container {
+    margin: 0;
+  }
+
+  .set-passcode-title,
+  .set-passcode-description {
+    text-align: center;
+  }
+
+  .forgot-your-passcode-placeholder {
+    height: 4.625rem;
+  }
+
+  @media screen and (min-height: 300px) and (max-height: 600px) {
+    .forgot-your-passcode-placeholder {
+      height: 4.5rem;
+    }
+  }
+}
